requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
python-socketio==5.8.0
gunicorn==21.2.0 